D2: use size_t instead of uint in IQueueInfo
May cause a very minor breakage but removes necessity in bunch
of casts. Most implementors of this method are in ocean itself
though.

+++ src/ocean/util/container/queue/NotifyingQueue.d
@@ -247,7 +247,7 @@ class NotifyingByteQueue : ISuspendable, IQueueInfo
 
     ***************************************************************************/
 
-    public uint length ( )
+    public size_t length ( )
     {
         return this.queue.length();
     }
